패스 경험치 및 레벨업 로직 1차 구현 - Redis Cluster에 유저별 각 패스 경험치와 레벨 저장 - 경험치 증가 요청시 경험치/레벨 상태 Redis에 반영 - 로그아웃 시 Redis의 각 패스 경험치/레벨을 MySQL에 업데이트하는 구조로 설계

+++ MySQLManager.cpp
@@ -465,7 +465,7 @@ bool MySQLManager::GetPassItemData(std::vector<std::pair<std::string, PassInfo>>
     }
 }
 
-bool MySQLManager::GetPassExpData(std::vector<uint32_t>& passExpLimit_) {
+bool MySQLManager::GetPassExpData(std::vector<uint16_t>& passExpLimit_) {
     semaphore.acquire();
 
     MYSQL* ConnPtr = GetConnection();
@@ -495,14 +495,14 @@ bool MySQLManager::GetPassExpData(std::vector<uint32_t>& passExpLimit_) {
             return false;
         }
 
-        std::unordered_map<uint16_t, uint32_t> passLevelExpMap;
+        std::unordered_map<uint16_t, uint16_t> passLevelExpMap;
         uint16_t maxLevel = 0;
 
         while ((Row = mysql_fetch_row(Result)) != NULL) {
             if (!Row[0] || !Row[1]) continue;
 
             uint16_t tempLevel = static_cast<uint16_t>(std::stoi(Row[0]));
-            uint32_t tempExp = static_cast<uint32_t>(std::stoi(Row[1]));
+            uint16_t tempExp = static_cast<uint16_t>(std::stoi(Row[1]));
 
             passLevelExpMap[tempLevel] = tempExp;
             if (tempLevel > maxLevel) maxLevel = tempLevel;
