+++ MySQLManager.h
@@ -27,7 +27,7 @@ public:
 	bool GetShopItemData(std::unordered_map<ShopItemKey, ShopItem, ShopItemKeyHash>& shopItemData_);
 	bool GetPassInfo(std::vector<std::pair<std::string, PassInfo>>& passInfoVector_);
 	bool GetPassItemData(std::vector<std::pair<std::string, PassInfo>>& passInfoVector_, std::unordered_map<std::string, std::unordered_map<PassDataKey, PassDataForSend, PassDataKeyHash>>& passDataMap_);
-	bool GetPassExpData(std::vector<uint32_t>& passExpLimit_);
+	bool GetPassExpData(std::vector<uint16_t>& passExpLimit_);
 
 
 	// ======================= SYNCRONIZATION =======================
